requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.1
PyMySQL==1.1.0
Flask-Mail==0.9.1
itsdangerous==2.1.2
cryptography==46.0.5